Extract user-facing sync/async flags to aioadapters
